set attribute expire time

+++ mount_slash/fidc_cli.h
@@ -71,7 +71,7 @@ struct fcmh_cli_info_dir {
 struct fcmh_cli_info {
 	struct sl_resm			*fci_resm;
 	long			 	 fci_expire;	/* attr expire time */
-	double			 	 fci_timeout;	/* FUSE attr timeout value */
+	long			 	 fci_timeout;	/* FUSE attr timeout value */
 
 	uint64_t                         fci_pino;	/* silly rename fields */
 	int                         	 fci_nopen;
